+++ lib/index.js
@@ -1,9 +1,7 @@
-/* global document */
 'use strict';
 
 const Data = require('./data');
-const Templates = require('./templates');
-const Utils = require('./utils');
+const Ui = require('./ui');
 
 if (!window.whoami) {
     window.alert('Are you on 10kft? And logged in?');
@@ -14,13 +12,9 @@ const userId = window.whoami.id;
 const displayFrom = new Date(Date.now() - 86400 * 31 * 1000);
 const displayTo = new Date(Date.now() + 86400 * 7 * 1000);
 
-Utils.existingTableFadeout();
-Data.load(userId, displayFrom, displayTo).then(({ timeEntries, projects, leaveTypes }) => {
+Ui.fadeout();
+Data.load(userId, displayFrom, displayTo).then((data) => {
 
-    Utils.existingTableRemove();
-
-    const html = Templates.table({ timeEntries, projects, leaveTypes });
-
-    document.querySelector('#personPageMainContentAreaTimeTracker')
-        .insertAdjacentHTML('beforebegin', html);
+    Ui.cleanup();
+    Ui.render(data);
 });
