+++ lib/templates/table.ejs
@@ -12,10 +12,9 @@
     </thead>
 
     <tbody class="tk-time-tracker">
-        <% Utils.forEach(timeEntries.data, (dateKey, dayData) => {
+        <% Utils.forEach(timeEntries.data, (dayData, dateKey) => {
 
-            const rowProjectData = dayData.projects;
-            const rowHours = Object.keys(rowProjectData).reduce((sum, projectKey) => sum + (rowProjectData[projectKey].hours || 0), 0);
+            const rowHours = Object.keys(dayData.projects).reduce((sum, projectKey) => sum + (dayData.projects[projectKey].hours || 0), 0);
         %>
 
         <% if (dayData.display) { %>
@@ -23,7 +22,7 @@
 
             <td style="white-space: nowrap;padding-right: 14px;"><%= dateKey %></td>
 
-            <% Utils.forEach(timeEntries.totals, (projectKey, projectData) => { %>
+            <% Object.keys(timeEntries.totals).forEach((projectKey) => { const projectData = dayData.projects[projectKey]; %>
             <td class="<%- Utils.getHourCellClass(projects, leaveTypes, projectKey, projectData) %>"><div class="tk-hours">
                 <%- Utils.getProjectHours(projectData, rowHours) %>
             </div></td>
