+++ lib/utils.js
@@ -1,4 +1,3 @@
-/* global document */
 'use strict';
 
 // avoid relying on npm :trollface:
@@ -7,6 +6,10 @@ exports.leftPad = (str, n) => '0'.repeat(n - ('' + str).length) + str;
 exports.isoDate = (date) => `${date.getFullYear()}-${exports.leftPad(date.getMonth() + 1, 2)}-${exports.leftPad(date.getDate(), 2)}`;
 
 exports.forEach = (object, callback) => {
+
+    if (!object) {
+        return;
+    }
 
     Object.keys(object).forEach((k) => callback(object[k], k));
 };
@@ -91,24 +94,4 @@ exports.getRowStyle = ({ weekday, today }) => {
     }
 
     return rowStyle;
-};
-
-exports.existingTableFadeout = function () {
-
-    const existing = document.getElementById('results-3048m');
-    if (!existing) {
-        return;
-    }
-
-    existing.style.opacity = '0.3';
-};
-
-exports.existingTableRemove = function () {
-
-    const existing = document.getElementById('results-3048m');
-    if (!existing) {
-        return;
-    }
-
-    existing.parentNode.removeChild(existing);
 };
